+++ src/test/java/CalendarDateTest.java
@@ -1,4 +1,4 @@
-package test;
+package test.java;
 
 import static org.junit.jupiter.api.Assertions.*;
