+++ src/test/java/ProjectImplTest.java
@@ -1,4 +1,4 @@
-package testing.memoranda;
+package test.java;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
