+++ src/test/java/SettingCalendarTest.java
@@ -1,4 +1,4 @@
-package testing.memoranda;
+package test.java;
 
 import static org.junit.jupiter.api.Assertions.*;
